+++ src/NetCore/Web/Api/Controllers/SubscribersController.cs
@@ -30,7 +30,7 @@ using E.Standard.Security.Cryptography.Abstractions;
 using E.Standard.Security.Cryptography.Services;
 using E.Standard.Security.Cryptography.Token;
 using E.Standard.Web.Abstractions;
-using E.Standard.WebApp.Attributes;
+using E.Standard.WebApp.Reflection;
 using E.Standard.WebGIS.Core;
 using E.Standard.WebGIS.Core.Models;
 using E.Standard.WebGIS.SDK.Services;