+++ src/NetStandard/E.Standard.WebApp/Abstraction/IEndPointReflectionProvider.cs
@@ -1,0 +1,10 @@
+﻿namespace E.Standard.WebApp.Abstraction;
+
+public interface IEndPointReflectionProvider
+{
+    T GetCustomAttribute<T>() where T : Attribute;
+
+    T GetControllerCustomAttribute<T>() where T : Attribute;
+
+    T GetActionMethodCustomAttribute<T>() where T : Attribute;
+}